requirements.txt
altair==5.5.0
astroid==3.0.2
attrs==23.1.0
autopep8==2.0.4
blinker==1.9.0
build==0.10.0
CacheControl==0.13.1
cachetools==5.5.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.8
crashtest==0.4.1
dill==0.3.7
distlib==0.3.7
docformatter==1.7.5
dulwich==0.21.6
et_xmlfile==2.0.0
filelock==3.12.4
gitdb==4.0.12
GitPython==3.1.44
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.3.0
Jinja2==3.1.3
jsonschema==4.17.3
keyring==24.2.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.6
narwhals==1.21.1
numpy==1.26.2
openpyxl==3.1.5
packaging==23.1
pandas==2.1.3
pexpect==4.8.0
pillow==11.1.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
protobuf==5.29.3
ptyprocess==0.7.0
pyarrow==18.1.0
pycodestyle==2.11.1
pycparser==2.21
pydeck==0.9.1
Pygments==2.19.1
pylint==3.0.3
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==1.2.0
pytz==2023.3.post1
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.3
six==1.16.0
smmap==5.0.2
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tomlkit==0.12.1
tornado==6.4.2
trove-classifiers==2023.9.19
typing_extensions==4.12.2
tzdata==2023.3
untokenize==0.1.1
urllib3==2.0.5
virtualenv==20.24.5
websockets==12.0
xattr==0.10.1
zipp==3.17.0
